Update user id to ulong to match the discord user id

+++ Noob.API/Models/UserCommand.cs
@@ -3,7 +3,7 @@ namespace Noob.API.Models
 {
     public class UserCommand
     {
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
         public int CommandId { get; set; }
         public DateTime ExecutedAt { get; set; }
     }
